Fix update issuer logic
Issue:

Discovery is getting stuck during few implementation

Root cause:

When updating the issuer of the certificate after each operation, the issuer update logic was wrong.

It was checking for only DN and if 2 certificates with same DN, it fails in JPA.

Fix:

The code is updated to get the list of certificates with the matching DN and perform signature verification to update the serial number

The file Certificate Service Implementation seems changed the whole file, This is because of the formatting of the file to properly formatted.

Changed lines are 265 to 275