Fix deletion of discovered certificates
Issue: When attempting to delete a certificate that is discovered, there is an exception that the code is trying to delete the entry from certificate content table before deleting it in the certificate table.

Fix: Added logic to handle the scenario. If the certificate is found to be discovered and if the discovery part of the certificate still exists, then delete only the content.. Else the code will delete the certificate table entry and then delete the certificate content, since the certificate table references the contents

closes #79